re-enable -Werror

+++ source/skyland/room.cpp
@@ -1108,8 +1108,8 @@ public:
 
         if (not countdown_) {
             if (pos.x.as_integer() + 8 >
-                pfrm.screen().get_view().int_center().x +
-                    pfrm.screen().size().x) {
+                (int)(pfrm.screen().get_view().int_center().x +
+                      pfrm.screen().size().x)) {
                 pos.x = pfrm.screen().get_view().int_center().x +
                         pfrm.screen().size().x - (16 + 4);
                 sprite_.set_flip({true, false});
@@ -1170,7 +1170,7 @@ void Room::finalize(Platform& pfrm, App& app)
              pfrm.screen().get_view().int_center().x + 8 -
                  (size().x * 16) / 2) or
             (visual_center().x.as_integer() - (size().x * 16) / 2 >
-             pfrm.screen().get_view().int_center().x + pfrm.screen().size().x);
+             (int)(pfrm.screen().get_view().int_center().x + pfrm.screen().size().x));
 
         const int t = debris_tile();
